+++ app/Models/orders_items_features.php
@@ -10,7 +10,7 @@ class orders_items_features extends Model
     use HasFactory;
 
     protected $casts = [
-        'price' => 'double',
+        'price' => 'float',
     ];
 
     protected $fillable = ['order_item_id','product_feature_id','price'];
